+++ CHANGELOG.md
@@ -5,7 +5,7 @@ Changelog for insanic
 0.7.12 (unreleased)
 -------------------
 
-- Nothing changed yet.
+- UPGRADE: increase log level to critical for uncaught exceptions
 
 
 0.7.11 (2019-05-29)